Move `ExecutionMode` and `CubeDim` to common, use only `cubecl_ir` in optimizer
