+++ crates/cubecl-common/Cargo.toml
@@ -14,7 +14,9 @@ repository = "https://github.com/tracel-ai/cubecl/tree/main/crates/cubecl-common
 version.workspace = true
 
 [features]
-default = ["std"]
+bitcode = ["dep:bitcode"]
+default = ["std", "serde"]
+serde = ["dep:serde"]
 std = ["rand/std", "futures-lite"]
 
 [target.'cfg(target_family = "wasm")'.dependencies]
@@ -23,6 +25,7 @@ web-time = { workspace = true }
 
 [dependencies]
 # ** Please make sure all dependencies support no_std when std is disabled **
+bitcode = { workspace = true, optional = true }
 bytemuck = { workspace = true, features = ["derive"] }
 derive-new = { workspace = true }
 derive_more = { workspace = true, features = [
@@ -37,7 +40,7 @@ half = { workspace = true }
 log = { workspace = true }
 num-traits = { workspace = true }
 rand = { workspace = true }
-serde = { workspace = true }
+serde = { workspace = true, optional = true }
 spin = { workspace = true } # using in place of use std::sync::Mutex;
 
 # Only activate futures for std env
