+++ crates/cubecl-common/src/lib.rs
@@ -27,6 +27,10 @@ pub mod future;
 
 extern crate alloc;
 
+/// Custom float implementations
 mod float;
+/// Common kernel types
+mod kernel;
 
 pub use float::*;
+pub use kernel::*;
