+++ crates/cubecl-core/src/codegen/execution.rs
@@ -1,8 +1,9 @@
-use crate::compute::{CubeDim, KernelTask};
+use crate::compute::KernelTask;
 use crate::frontend::TensorHandleRef;
 use crate::ir::Elem;
 use crate::pod::CubeElement;
 use crate::{calculate_cube_count_elemwise, Kernel, Runtime};
+use cubecl_common::CubeDim;
 use cubecl_runtime::client::ComputeClient;
 use cubecl_runtime::server::{Binding, CubeCount, Handle};
