+++ crates/cubecl-opt/Cargo.toml
@@ -11,12 +11,13 @@ repository = "https://github.com/tracel-ai/cubecl/tree/main/cubecl-opt"
 version.workspace = true
 
 [features]
-default = ["std", "cubecl-common/default", "cubecl-core/default"]
-std = ["cubecl-common/std", "cubecl-core/std"]
+default = ["std", "cubecl-common/default", "cubecl-ir/default"]
+std = ["cubecl-common/std"]
 
 [dependencies]
 cubecl-common = { path = "../cubecl-common", version = "0.5.0", default-features = false }
-cubecl-core = { path = "../cubecl-core", version = "0.5.0", default-features = false }
+cubecl-ir = { path = "../cubecl-ir", version = "0.5.0", default-features = false }
+
 float-ord = "0.3"
 log = "0.4"
 num = "0.4"
@@ -24,3 +25,6 @@ petgraph = { version = "0.6" }
 smallvec = { version = "1", features = ["union", "const_generics"] }
 stable-vec = { version = "0.4" }
 type-map = { version = "0.5" }
+
+[dev-dependencies]
+cubecl-core = { path = "../cubecl-core", version = "0.5.0", default-features = false }
