+++ crates/cubecl-opt/src/passes/index_merge.rs
@@ -1,8 +1,6 @@
 use std::collections::{HashMap, HashSet};
 
-use cubecl_core::ir::{
-    CopyMemoryOperator, Id, Instruction, Operation, Operator, Variable, VariableKind,
-};
+use cubecl_ir::{CopyMemoryOperator, Id, Instruction, Operation, Operator, Variable, VariableKind};
 
 use crate::{AtomicCounter, Optimizer};
