+++ crates/cubecl-opt/src/passes/inlined_if_to_select.rs
@@ -1,6 +1,6 @@
 use std::mem::take;
 
-use cubecl_core::ir::{Instruction, Operator, Select};
+use cubecl_ir::{Instruction, Operator, Select};
 use petgraph::{graph::NodeIndex, visit::EdgeRef};
 
 use crate::{passes::update_references, AtomicCounter, ControlFlow, Optimizer};
