+++ crates/cubecl-wgpu/src/compiler/spirv.rs
@@ -7,6 +7,7 @@ use ash::{
         TRUE,
     },
 };
+use cubecl_common::ExecutionMode;
 use cubecl_core::{
     channel::MutexComputeChannel,
     client::ComputeClient,
@@ -14,7 +15,7 @@ use cubecl_core::{
     ir::{Elem, FloatKind, IntKind, UIntKind},
     prelude::CompiledKernel,
     server::ComputeServer,
-    AtomicFeature, ExecutionMode, Feature, Runtime,
+    AtomicFeature, Feature, Runtime,
 };
 use cubecl_runtime::{ComputeRuntime, DeviceProperties};
 use cubecl_spirv::CompilationOptions;
